move exception handler related code to trait
