+++ src/Concerns/RegistersExceptionHandlers.php
@@ -1,0 +1,96 @@
+<?php
+
+namespace Spatie\FlareClient\Concerns;
+
+use ErrorException;
+use Spatie\FlareClient\Flare;
+use Throwable;
+
+/** @mixin Flare */
+trait RegistersExceptionHandlers
+{
+    /** @var null|callable */
+    protected $previousExceptionHandler = null;
+
+    /** @var null|callable */
+    protected $previousErrorHandler = null;
+
+    protected static ?string $memoryReserve = null;
+
+    public function registerFlareHandlers(): self
+    {
+        $this->registerExceptionHandler();
+
+        $this->registerErrorHandler();
+
+        $this->registerShutdownHandler();
+
+        return $this;
+    }
+
+    public function registerExceptionHandler(): self
+    {
+        $this->previousExceptionHandler = set_exception_handler(fn ($e) => $this->handleException($e));
+
+        return $this;
+    }
+
+    public function registerErrorHandler(): self
+    {
+        set_error_handler(function ($level, $message, $file = '', $line = 0) {
+            $this->handleError($level, $message, $file, $line);
+        });
+
+        return $this;
+    }
+
+    public function registerShutdownHandler(): self
+    {
+        register_shutdown_function(fn () => $this->handleShutdown());
+
+        return $this;
+    }
+
+    public function handleException(Throwable $throwable): void
+    {
+        $this->report($throwable);
+
+        if ($this->previousExceptionHandler
+            && is_callable($this->previousExceptionHandler)) {
+            call_user_func($this->previousExceptionHandler, $throwable);
+        }
+    }
+
+    /**
+     * @return mixed
+     */
+    public function handleError(mixed $code, string $message, string $file = '', int $line = 0)
+    {
+        $exception = new ErrorException($message, 0, $code, $file, $line);
+
+        $this->report($exception);
+
+        if ($this->previousErrorHandler) {
+            return call_user_func(
+                $this->previousErrorHandler,
+                $message,
+                $code,
+                $file,
+                $line
+            );
+        }
+    }
+
+    protected function handleShutdown(): void
+    {
+        self::$memoryReserve = null;
+
+        $error = error_get_last(); // @todo returns null when running from a test file
+
+        if ($error !== null && $error['type'] === E_ERROR) {
+            ini_set('memory_limit', '-1'); // @todo perhaps something else as -1 is better
+
+            $this->handleError($error['type'], $error['message'], $error['file'], $error['line']);
+        }
+    }
+}
